#127 - Fix: components names
